requirements.txt
flask
openai
pillow
gunicorn
